chore: move get_verified_packed_client_and_proof_update out of ckb-chain module
