+++ .gitignore
@@ -27,3 +27,6 @@ mc.log
 
 # Ignore OSX .DS_Store file
 .DS_Store
+
+# Ignore generated rocksdb storage folder
+test_storage/
