+++ crates/relayer/src/chain/ckb.rs
@@ -1,23 +1,14 @@
 use ckb_jsonrpc_types::OutputsValidator;
 use ckb_sdk::{Address, AddressPayload, NetworkType};
 use eth2_types::MainnetEthSpec;
-use eth_light_client_in_ckb_verification::{
-    mmr,
-    types::{
-        core::{Client as EthLcClient, Header as EthLcHeader},
-        packed::{self, Client as PackedClient, ProofUpdate as PackedProofUpdate},
-        prelude::*,
-    },
-};
-use ibc_relayer_storage::{error::Error as StorageError, prelude::*, Storage};
+use ibc_relayer_storage::Storage;
 use ibc_relayer_types::clients::ics07_ckb::{
     client_state::ClientState as CkbClientState,
     consensus_state::ConsensusState as CkbConsensusState, header::Header as CkbHeader,
     light_block::LightBlock as CkbLightBlock,
 };
 use ibc_relayer_types::clients::ics07_eth::{
-    client_state::ClientState as EthClient,
-    types::{Header as EthHeader, Update as EthUpdate},
+    client_state::ClientState as EthClient, types::Update as EthUpdate,
 };
 use ibc_relayer_types::{
     core::{
@@ -35,7 +26,6 @@ use ibc_relayer_types::{
 };
 use semver::Version;
 use std::sync::{Arc, RwLock};
-use tendermint_light_client::errors::Error as LightClientError;
 use tendermint_rpc::endpoint::broadcast::tx_sync::Response;
 use tokio::runtime::Runtime as TokioRuntime;
 
@@ -76,6 +66,7 @@ mod assembler;
 mod communication;
 mod helper;
 mod signer;
+mod utils;
 
 #[cfg(test)]
 mod mock_rpc_client;
@@ -125,8 +116,17 @@ impl CkbChain {
         &mut self,
         header_updates: Vec<EthUpdate>,
     ) -> Result<Vec<IbcEventWithHeight>, Error> {
+        let onchain_packed_client_opt = self.rt.block_on(self.rpc_client.fetch_packed_client(
+            &self.config.lightclient_contract_typeargs,
+            &self.config.id.to_string(),
+        ))?;
         let (packed_client, packed_proof_update) =
-            self.get_verified_packed_client_and_proof_update(header_updates)?;
+            utils::get_verified_packed_client_and_proof_update(
+                &self.id().to_string(),
+                header_updates,
+                &self.storage,
+                onchain_packed_client_opt,
+            )?;
 
         let tx_assembler_address = self.tx_assembler_address()?;
         let (tx, inputs) = self
@@ -151,165 +151,6 @@ impl CkbChain {
             .map_err(|e| Error::rpc_response(e.to_string()))?;
 
         Ok(vec![])
-    }
-
-    pub(self) fn get_verified_packed_client_and_proof_update(
-        &self,
-        header_updates: Vec<EthUpdate>,
-    ) -> Result<(PackedClient, PackedProofUpdate), Error> {
-        if header_updates.is_empty() {
-            return Err(Error::empty_upgraded_client_state());
-        }
-        let start_slot = header_updates[0].finalized_header.slot;
-        for (i, item) in header_updates.iter().enumerate() {
-            if item.finalized_header.slot != i as u64 + start_slot {
-                return Err(Error::send_tx("uncontinuous header slot".to_owned()));
-            }
-        }
-
-        let mut is_creation = true;
-        // Check the tip in storage and the tip in the client cell are the same.
-        if let Some(stored_tip_slot) = self.storage.get_tip_beacon_header_slot()? {
-            if start_slot != stored_tip_slot + 1 {
-                let height = (stored_tip_slot + 1).try_into().expect("slot too big");
-                return Err(Error::light_client_verification(
-                    self.id().to_string(),
-                    LightClientError::missing_last_block_id(height),
-                ));
-            }
-            is_creation = false;
-        }
-        let onchain_packed_client_opt = self.rt.block_on(self.rpc_client.fetch_packed_client(
-            &self.config.lightclient_contract_typeargs,
-            &self.config.id.to_string(),
-        ))?;
-
-        if is_creation != onchain_packed_client_opt.is_none() {
-            return Err(Error::send_tx(
-                "storage and the chain state is not same".to_owned(),
-            ));
-        }
-
-        if let Some(ref client) = onchain_packed_client_opt {
-            let onchain_tip_slot: u64 = client.maximal_slot().unpack();
-            if start_slot != onchain_tip_slot + 1 {
-                let height = (onchain_tip_slot + 1).try_into().expect("slot too big");
-                return Err(Error::light_client_verification(
-                    self.id().to_string(),
-                    LightClientError::missing_last_block_id(height),
-                ));
-            }
-        }
-
-        let finalized_headers = header_updates
-            .iter()
-            .map(|update| {
-                let EthHeader {
-                    slot,
-                    proposer_index,
-                    parent_root,
-                    state_root,
-                    body_root,
-                } = update.finalized_header.clone();
-                let header = EthLcHeader {
-                    slot,
-                    proposer_index,
-                    parent_root,
-                    state_root,
-                    body_root,
-                };
-                header.calc_cache()
-            })
-            .collect::<Vec<_>>();
-
-        let minimal_slot = self
-            .storage
-            .get_base_beacon_header_slot()?
-            .unwrap_or(start_slot);
-        let last_finalized_header = &finalized_headers[finalized_headers.len() - 1];
-        let maximal_slot = last_finalized_header.inner.slot;
-
-        // Saves all header digests into storage for MMR.
-        {
-            let mut finalized_headers_iter = finalized_headers.iter();
-
-            let mut last_slot = if self.storage.is_initialized()? {
-                start_slot - 1
-            } else {
-                let first = finalized_headers_iter.next().expect("checked");
-                self.storage
-                    .initialize_with(first.inner.slot, first.digest())?;
-                self.storage.put_tip_beacon_header_slot(first.inner.slot)?;
-                first.inner.slot
-            };
-
-            let mut mmr = self.storage.chain_root_mmr(last_slot)?;
-
-            for header in finalized_headers_iter {
-                last_slot = header.inner.slot;
-                mmr.push(header.digest()).map_err(StorageError::from)?;
-            }
-            mmr.commit().map_err(StorageError::from)?;
-
-            self.storage.put_tip_beacon_header_slot(last_slot)?;
-        };
-
-        // Gets the new root and a proof for all new headers.
-        let (packed_headers_mmr_root, packed_headers_mmr_proof) = {
-            let positions = (start_slot..=maximal_slot)
-                .into_iter()
-                .map(|slot| mmr::lib::leaf_index_to_pos(slot - minimal_slot))
-                .collect::<Vec<_>>();
-
-            let mmr = self.storage.chain_root_mmr(maximal_slot)?;
-
-            let headers_mmr_root = mmr.get_root().map_err(StorageError::from)?;
-            let headers_mmr_proof_items = mmr
-                .gen_proof(positions)
-                .map_err(StorageError::from)?
-                .proof_items()
-                .iter()
-                .map(Clone::clone)
-                .collect::<Vec<_>>();
-            let headers_mmr_proof = packed::MmrProof::new_builder()
-                .set(headers_mmr_proof_items)
-                .build();
-
-            (headers_mmr_root, headers_mmr_proof)
-        };
-
-        // Build the packed proof update.
-        let packed_proof_update = {
-            let updates_items = finalized_headers
-                .iter()
-                .map(|header| {
-                    packed::FinalityUpdate::new_builder()
-                        .finalized_header(header.inner.pack())
-                        .build()
-                })
-                .collect::<Vec<_>>();
-            let updates = packed::FinalityUpdateVec::new_builder()
-                .set(updates_items)
-                .build();
-            packed::ProofUpdate::new_builder()
-                .new_headers_mmr_root(packed_headers_mmr_root)
-                .new_headers_mmr_proof(packed_headers_mmr_proof)
-                .updates(updates)
-                .build()
-        };
-
-        // Invoke verification from core::Client on packed_proof_update
-        let client = if let Some(client) = onchain_packed_client_opt {
-            client
-                .unpack()
-                .try_apply_packed_proof_update(packed_proof_update.as_reader())
-                .map_err(|_| Error::send_tx("failed to update header".to_owned()))?
-        } else {
-            EthLcClient::new_from_packed_proof_update(packed_proof_update.as_reader())
-                .map_err(|_| Error::send_tx("failed to create header".to_owned()))?
-        };
-
-        Ok((client.pack(), packed_proof_update))
     }
 
     pub fn network(&self) -> Result<NetworkType, Error> {
